Move reputation to utils
